LayoutTests/imported/w3c/web-platform-tests/content-security-policy/style-src/style-src-injected-stylesheet-allowed.sub-expected.txt


FAIL Programatically injected stylesheet should load assert_equals: expected "2px" but got "0px"
Lorem ipsum
